rename driver to IO for adaravis and kinetix
