+++ src/ophyd_async/epics/adkinetix/__init__.py
@@ -1,9 +1,9 @@
 from ._kinetix import KinetixDetector
 from ._kinetix_controller import KinetixController
-from ._kinetix_driver import KinetixDriver
+from ._kinetix_io import KinetixDriverIO
 
 __all__ = [
     "KinetixDetector",
     "KinetixController",
-    "KinetixDriver",
+    "KinetixDriverIO",
 ]
