+++ src/ophyd_async/epics/adkinetix/_kinetix_controller.py
@@ -4,7 +4,7 @@ from typing import Optional
 from ophyd_async.core import AsyncStatus, DetectorControl, DetectorTrigger
 from ophyd_async.epics import adcore
 
-from ._kinetix_driver import KinetixDriver, KinetixTriggerMode
+from ._kinetix_io import KinetixDriverIO, KinetixTriggerMode
 
 KINETIX_TRIGGER_MODE_MAP = {
     DetectorTrigger.internal: KinetixTriggerMode.internal,
@@ -17,7 +17,7 @@ KINETIX_TRIGGER_MODE_MAP = {
 class KinetixController(DetectorControl):
     def __init__(
         self,
-        driver: KinetixDriver,
+        driver: KinetixDriverIO,
     ) -> None:
         self._drv = driver
