+++ src/ophyd_async/epics/adkinetix/_kinetix_io.py
@@ -16,7 +16,7 @@ class KinetixReadoutMode(str, Enum):
     dynamic_range = 3
 
 
-class KinetixDriver(adcore.ADBase):
+class KinetixDriverIO(adcore.ADBase):
     def __init__(self, prefix: str, name: str = "") -> None:
         # self.pixel_format = epics_signal_rw_rbv(PixelFormat, prefix + "PixelFormat")
         self.trigger_mode = epics_signal_rw_rbv(
